declare ButtonType enum outside LargeButton's file
