+++ lib/ui/components/models/button_type.dart
@@ -1,0 +1,10 @@
+enum ButtonType {
+  primary,
+  secondary,
+  neutral,
+  trailingIcon,
+  critical,
+  tertiaryCritical,
+  trailingIconPrimary,
+  trailingIconSecondary,
+}
